third_party/blink/web_tests/external/wpt/mathml/presentation-markup/fractions/frac-parameters-1-expected.txt

This is a testharness.js-based test.
PASS AxisHeight
PASS DenominatorDisplayStyleGapMin
PASS DenominatorDisplayStyleShiftDown
FAIL DenominatorGapMin assert_approx_equals: mfrac: denominator gap expected 40 +/- 1 but got -0.015625
FAIL DenominatorShiftDown assert_approx_equals: mfrac: denominator shift expected 30 +/- 1 but got 0
PASS NumeratorDisplayStyleGapMin
PASS NumeratorDisplayStyleShiftDown
FAIL NumeratorGapMin assert_approx_equals: mfrac: numerator gap expected 90 +/- 1 but got -0.015625
FAIL NumeratorShiftDown assert_approx_equals: mfrac: numerator shift expected 110 +/- 1 but got 14.96875
PASS FractionRuleThickness
Harness: the test ran to completion.
